Now merging in TestSelection only.
