+++ surefire-provider/src/test/java/org/arquillian/smart/testing/surefire/provider/TestStrategyApplierTest.java
@@ -1,162 +1,59 @@
 package org.arquillian.smart.testing.surefire.provider;
 
-import java.util.HashSet;
-import java.util.LinkedHashSet;
-import java.util.Set;
-import net.jcip.annotations.NotThreadSafe;
-import org.apache.maven.surefire.providerapi.ProviderParameters;
+import java.util.Collection;
 import org.apache.maven.surefire.util.TestsToRun;
-import org.arquillian.smart.testing.Configuration;
 import org.arquillian.smart.testing.TestSelection;
-import org.arquillian.smart.testing.spi.TestExecutionPlanner;
-import org.junit.Rule;
+import org.assertj.core.api.Assertions;
 import org.junit.Test;
-import org.junit.contrib.java.lang.system.RestoreSystemProperties;
-import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.when;
+import static java.util.Arrays.asList;
 
 @RunWith(MockitoJUnitRunner.class)
-@Category(NotThreadSafe.class)
 public class TestStrategyApplierTest {
 
-    @Rule
-    public final RestoreSystemProperties restoreSystemProperties = new RestoreSystemProperties();
+    @Mock
+    private TestsToRun testsToRun;
 
     @Mock
     private TestExecutionPlannerLoader testExecutionPlannerLoader;
 
-    @Mock
-    private ProviderParameters providerParameters;
-
-    @Mock
-    private TestExecutionPlanner testExecutionPlanner;
-
     @Test
-    public void should_return_tests_only_relevant_for_defined_strategies_when_selecting_mode_configured() {
+    public void should_return_merged_test_selections_if_test_selection_has_same_class() {
         // given
-        System.setProperty(Configuration.SMART_TESTING_MODE, "selecting");
-        System.setProperty(Configuration.SMART_TESTING, "static");
-
-        final Set<Class<?>> defaultTestsToRun = new HashSet<>();
-        defaultTestsToRun.add(TestNgParametersTest.class);
-        defaultTestsToRun.add(TestExecutionPlannerLoaderTest.class);
-
-        final TestsToRun testsToRun = new TestsToRun(defaultTestsToRun);
-        when(testExecutionPlannerLoader.getPlannerForStrategy("static")).thenReturn(testExecutionPlanner);
-        when(providerParameters.getTestClassLoader()).thenReturn(Thread.currentThread().getContextClassLoader());
-
-        final Set<TestSelection> strategyTests = new LinkedHashSet<>();
-        strategyTests.add(new TestSelection(TestExecutionPlannerLoaderTest.class.getName(), "static"));
-
-        when(testExecutionPlanner.getTests()).thenReturn(strategyTests);
+        final TestSelection
+            testSelectionNew = new TestSelection(TestExecutionPlannerLoaderTest.class.getName(), "new");
+        final TestSelection testSelectionChanged = new TestSelection(TestExecutionPlannerLoaderTest.class.getName(), "changed");
+        final TestStrategyApplier testStrategyApplier = new TestStrategyApplier(testsToRun, testExecutionPlannerLoader, Thread.currentThread().getContextClassLoader());
 
         // when
-        final Configuration configuration = Configuration.load();
-        TestStrategyApplier testStrategyApplier = new TestStrategyApplier(testsToRun, testExecutionPlannerLoader, providerParameters.getTestClassLoader());
-        final TestsToRun realTestPlanning = testStrategyApplier.apply(configuration);
+        final Collection<TestSelection> testSelections =
+            testStrategyApplier.filterAndMergeTestSelection(asList(testSelectionNew, testSelectionChanged));
 
-        // then
-        assertThat(realTestPlanning.getLocatedClasses())
+        //then
+        Assertions.assertThat(testSelections)
             .hasSize(1)
-            .containsExactly(TestExecutionPlannerLoaderTest.class);
-
+            .flatExtracting("types").containsExactly("new", "changed");
     }
 
     @Test
-    public void should_return_tests_selected_by_default() {
+    public void should_not_return_merged_test_selections_if_test_selection_has_different_class() {
         // given
-        System.setProperty(Configuration.SMART_TESTING, "static");
-
-        final Set<Class<?>> defaultTestsToRun = new LinkedHashSet<>();
-        defaultTestsToRun.add(TestNgParametersTest.class);
-        defaultTestsToRun.add(TestStrategyApplierTest.class);
-        defaultTestsToRun.add(TestExecutionPlannerLoaderTest.class);
-
-        final TestsToRun testsToRun = new TestsToRun(defaultTestsToRun);
-
-        when(testExecutionPlannerLoader.getPlannerForStrategy("static")).thenReturn(testExecutionPlanner);
-
-        final Set<TestSelection> strategyTests = new LinkedHashSet<>();
-        strategyTests.add(new TestSelection(TestExecutionPlannerLoaderTest.class.getName(), "static"));
-
-        when(testExecutionPlanner.getTests()).thenReturn(strategyTests);
+        final TestSelection testSelectionNew = new TestSelection(TestExecutionPlannerLoaderTest.class.getName(), "new");
+        final TestSelection testSelectionChanged =
+            new TestSelection(TestStrategyApplierTest.class.getName(), "changed");
+        final TestStrategyApplier testStrategyApplier = new TestStrategyApplier(testsToRun, testExecutionPlannerLoader, Thread.currentThread().getContextClassLoader());
 
         // when
-        final Configuration configuration = Configuration.load();
-        TestStrategyApplier testStrategyApplier = new TestStrategyApplier(testsToRun, testExecutionPlannerLoader,
-            Thread.currentThread().getContextClassLoader());
-        final TestsToRun realTestPlanning = testStrategyApplier.apply(configuration);
+        final Collection<TestSelection> testSelections =
+            testStrategyApplier.filterAndMergeTestSelection(asList(testSelectionNew, testSelectionChanged));
+
 
         // then
-        assertThat(realTestPlanning.getLocatedClasses())
-            .containsExactly(TestExecutionPlannerLoaderTest.class);
-    }
-
-    @Test
-    public void should_return_all_tests_when_ordering_selected() {
-        // given
-        System.setProperty(Configuration.SMART_TESTING_MODE, "ordering");
-        System.setProperty(Configuration.SMART_TESTING, "static");
-
-        final Set<Class<?>> defaultTestsToRun = new LinkedHashSet<>();
-        defaultTestsToRun.add(TestNgParametersTest.class);
-        defaultTestsToRun.add(TestStrategyApplierTest.class);
-        defaultTestsToRun.add(TestExecutionPlannerLoaderTest.class);
-
-        final TestsToRun testsToRun = new TestsToRun(defaultTestsToRun);
-
-        when(testExecutionPlannerLoader.getPlannerForStrategy("static")).thenReturn(testExecutionPlanner);
-
-        final Set<TestSelection> strategyTests = new LinkedHashSet<>();
-        strategyTests.add(new TestSelection(TestExecutionPlannerLoaderTest.class.getName(), "static"));
-
-        when(testExecutionPlanner.getTests()).thenReturn(strategyTests);
-
-        // when
-        final Configuration configuration = Configuration.load();
-        TestStrategyApplier testStrategyApplier = new TestStrategyApplier(testsToRun, testExecutionPlannerLoader,
-            Thread.currentThread().getContextClassLoader());
-        final TestsToRun realTestPlanning = testStrategyApplier.apply(configuration);
-
-        // then
-        assertThat(realTestPlanning.getLocatedClasses())
-            .containsExactly(TestExecutionPlannerLoaderTest.class, TestNgParametersTest.class, TestStrategyApplierTest.class);
-    }
-
-    @Test
-    public void should_not_return_test_from_strategies_if_it_is_not_in_class_path() {
-        // given
-        System.setProperty(Configuration.SMART_TESTING_MODE, "selecting");
-        System.setProperty(Configuration.SMART_TESTING, "static");
-
-        final Set<Class<?>> defaultTestsToRun = new HashSet<>();
-        defaultTestsToRun.add(TestNgParametersTest.class);
-        defaultTestsToRun.add(TestExecutionPlannerLoaderTest.class);
-
-        final TestsToRun testsToRun = new TestsToRun(defaultTestsToRun);
-        when(testExecutionPlannerLoader.getPlannerForStrategy("static")).thenReturn(testExecutionPlanner);
-
-        final Set<TestSelection> strategyTests = new LinkedHashSet<>();
-        strategyTests.add(new TestSelection(TestExecutionPlannerLoaderTest.class.getName(), "static"));
-        strategyTests.add(new TestSelection("org.arquillian.smart.testing.vcs.git.ChangedFilesDetectorTest", "static"));
-
-        when(testExecutionPlanner.getTests()).thenReturn(strategyTests);
-
-        // when
-        final Configuration configuration = Configuration.load();
-        TestStrategyApplier testStrategyApplier = new TestStrategyApplier(testsToRun, testExecutionPlannerLoader, Thread.currentThread().getContextClassLoader());
-        final TestsToRun realTestPlanning = testStrategyApplier.apply(configuration);
-
-        // then
-
-        assertThat(realTestPlanning.getLocatedClasses())
-            .hasSize(1)
-            .containsExactly(TestExecutionPlannerLoaderTest.class);
-
+        Assertions.assertThat(testSelections)
+            .hasSize(2)
+            .flatExtracting("types").containsExactly("new", "changed");
     }
 }
